backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AI Integration
anthropic==0.18.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2

# Task Management Integrations
py-trello==0.19.0
jira==3.5.2

# Email & Scheduling
fastapi-mail==1.4.1
apscheduler==3.10.4

# Data Processing
pandas==2.2.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utils
python-dateutil==2.8.2
pytz==2024.1
